+++ HISTORY.md
@@ -1,9 +1,12 @@
 # Snap! (BYOB) History
 
 ## in development:
+* **Notable Changes:**
+    * deprecated "expand (input) to (n) slots" command, because "set slot ... to ..." can now do this by passing a list as value
 
 ### 2024-11-09
 * threads: enhanced "set slot ... to ..." command to also handle variadic inputs, setting them to a list of values
+* threads, objects: deprecated "expand (input) to (n) slots" command, because "set slot ... to ..." can now do this by passing a list as value
 
 ## 10.2.0:
 * **New Features:**